Huge update with new commands and revamps

+++ modules/admin.py
@@ -1,10 +1,11 @@
 """Definition of the bot's Admin module.'"""
+from __future__ import print_function
 import subprocess
 
 import discord
 from discord.ext import commands
+from util.perms import check_perms
 from .cog import Cog
-
 
 class Admin(Cog):
     """Commands useful for admins and/or moderators.
@@ -15,30 +16,37 @@ class Admin(Cog):
         """Checks if author of a message is this bot."""
         return mem.author == self.bot.user
 
-    @commands.command()
-    async def purge(self, channel: discord.Channel):
+    async def perm_err(self, ctx):
+        """Returns a permission error in the chat."""
+        await self.bot.say(ctx.message.author.mention + ' You don\'t have enough **permissions** to execute that command!')
+
+    @commands.command(pass_context=True)
+    async def purge(self, ctx, channel: discord.Channel):
         """Removes all of this bot's messages on a channel.
         Syntax: purge [channel name]"""
+        if not check_perms(ctx, ['bot_owner']):
+            await self.perm_err(ctx)
+            return
         deleted = await self.bot.purge_from(channel, limit=200, check=self.is_me)
         await self.bot.send_message(channel, 'Deleted {} message(s)'.format(len(deleted)))
 
-    @commands.command()
-    async def nuke(self, channel: discord.Channel):
+    @commands.command(pass_context=True)
+    async def nuke(self, ctx, channel: discord.Channel):
         """NUKES a channel by deleting all messages!
         Syntax: nuke [channel name]"""
+        if not check_perms(ctx, ['bot_owner']):
+            await self.perm_err(ctx)
+            return
         deleted = await self.bot.purge_from(channel, limit=1000)
         await self.bot.send_message(channel, 'Deleted {} message(s)'.format(len(deleted)))
-
-    @commands.command()
-    async def say(self, *args):
-        """Simply sends the input as a message. For testing.
-        Syntax: say [message]"""
-        await self.bot.say(' '.join(args))
 
     @commands.command(pass_context=True)
     async def update(self, ctx):
         """Auto-updates this bot and restarts if any code was updated.
         Syntax: update"""
+        if not check_perms(ctx, ['bot_owner']):
+            await self.perm_err(ctx)
+            return
         await self.bot.say('Trying to update...')
         try:
             gitout = subprocess.check_output(['git', 'pull', '-v'], stderr=subprocess.STDOUT).decode('utf-8')
@@ -57,9 +65,16 @@ class Admin(Cog):
             await self.bot.say('Bot was able to update, now restarting.')
             ctx.invoke(self.restart)
 
-    @commands.command()
-    async def restart(self):
+    @commands.command(pass_context=True)
+    async def restart(self, ctx):
         """Restarts this bot.
         Syntax: restart"""
-        await self.bot.say('This bot is now restarting!')
-        exit(0)
+        if not check_perms(ctx, ['bot_owner']):
+            await self.perm_err(ctx)
+            return
+#        for i in self.bot.servers:
+#            await self.bot.send_message(i.default_channel, 'This bot (' + self.bname + ') is now restarting!')
+        await self.bot.say('This bot (' + self.bname + ') is now restarting!')
+        print('This bot is now restarting!')
+        self.bot.is_restart = True
+        self.loop.stop()
